Keep shared logic of query in superclass
Also fixed some javadoc and typing
